+++ rgbkbd/rgbkbd_daemon.cc
@@ -38,7 +38,7 @@ void DBusAdaptor::SetCapsLockState(bool enabled) {
   rgb_keyboard_controller_.SetCapsLockState(enabled);
 }
 
-void DBusAdaptor::SetStaticBackgroundColor(uint32_t r, uint32_t g, uint32_t b) {
+void DBusAdaptor::SetStaticBackgroundColor(uint8_t r, uint8_t g, uint8_t b) {
   rgb_keyboard_controller_.SetStaticBackgroundColor(r, g, b);
 }
